Provide full chain in ingress TLS certificate
The terraform data source returns the chain in the reverse order. We
initially stripped the chain and kept only the end user certificate. It
is accepted by most browsers but is flagged as not compliant by palo
alto.
We now provide the full chain, in order compliant with the TLS RFC.