Enable welcome app
The welcome app is not related to business but is required on top
domains such as test.teacherservices.cloud. A trusted website must exist
on the top domain so it is entered in the Palo Alto database and the
subdomains are trusted.
This solves the security errors when users are on the the DfE network.